Add a cool test
One TODO entry concerned turning the .crossrefs array that is parsed out
of the OEIS record into a lazy Seq that returns OEIS::Entry objects for
the referenced sequences. This was *not* implemented because the field
contains textual information (explanations, etc.) than just sequence IDs.

The implementation made it into the 02-attributes.t test.