Refactor OEIS::lookup and OEIS::Entry creation
The way from OEIS::lookup to OEIS::Entry objects is now divided more
neatly into three stages.

The first stage is carried out by the new `our` OEIS::fetch routine
which replaces the inelegant lookup'multi. It delegates the actual
fetching of plain OEIS pages to fetch'paginate. To do pagination,
that sub has to examine the header of each page. It already throws
some exceptions it can see from the header.

The pages are then passed to chop-record which transforms a Seq of
pages into a Seq of records. A record consists of just the %-prefixed
"tag lines" that hold the OEIS payload.

The parsing of individual tag lines still remains in OEIS::Entry,
but the constructor was renamed to just .parse.

This new division is: cleaner, allows to intercept the raw OEIS
pages, store them and read them back into OEIS::Entry objects,
which will come in handy soon for offline testing, and as a by-
product improves fetch'paginate in that it doesn't have to fetch
the last page twice to notice that it has reached the end.